+++ CHANGELOG.md
@@ -45,6 +45,7 @@ returned an empty array, `getFillable` would be used instead.
 - The `Schema\EloquentSchema` is deprecated in favour of using the `Eloquent\AbstractSchema`.
 - The `Store\EloquentAdapter` is deprecated in favour of using the `Eloquent\AbstractAdapter`.
 - The `Testing\InteractsWithModels` trait is deprecated in favour of Laravel database assertion helpers.
+- The `ErrorBag::toArray` method is deprecated in favour of `ErrorBag::all`.
 
 ### Fixed
 - [#128](https://github.com/cloudcreativity/laravel-json-api/issues/128)